% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censcast-package.R
\docType{package}
\name{censcast-package}
\alias{censcast}
\alias{censcast-package}
\title{censcast: hospital census forecasts from hubverse admission forecasts}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/cyrilgeismar/censcast}
\item Report bugs at \url{https://github.com/cyrilgeismar/censcast/issues}
}
}
\author{
\strong{Maintainer}: Cyril Geismar \email{cyrilgeismar@gmail.com}
}
\keyword{internal}
